minor bugs

+++ kiwix-zim-updater.sh
@@ -96,7 +96,7 @@ master_scrape() {
 
   FoundURLs=$(echo "$BaseURL" | wc -l)
 
-  if [[ $FoundURLs -eq 0 ]]; then
+  if [[ -z "$BaseURL" ]]; then
     echo -e "${RED_REGULAR}  ✗ No ZIM URLs found. Exiting...${CLEAR}"
     exit 0
   elif [[ $FoundURLs -ne 1 ]]; then
@@ -806,8 +806,8 @@ if [ $AnyDownloads -eq 1 ]; then
     if [[ $DownloadFailed -eq 1 ]] || [[ $SKIP_PURGE -eq 1 ]] || [[ $VERIFY_LIBRARY -eq 1 ]]; then
       [[ $DEBUG -eq 0 ]] && echo "End : $(date -u)" >>download.log
       [[ $DEBUG -eq 1 ]] && echo "End : $(date -u) *** Simulation ***" >>download.log
-      [[ $SKIP_PURGE -eq 1 ]] && [[ $DownloadFailed -ne 1 ]] && rm "$LockFilePath"
-      [[ $VERIFY_LIBRARY -eq 1 ]] && [[ $RequiresDownload -eq 1 ]] && rm "$LockFilePath"
+      [[ $DEBUG -eq 0 ]] && [[ $SKIP_PURGE -eq 1 ]] && [[ $DownloadFailed -ne 1 ]] && rm "$LockFilePath"
+      [[ $DEBUG -eq 0 ]] && [[ $VERIFY_LIBRARY -eq 1 ]] && [[ $RequiresDownload -eq 1 ]] && rm "$LockFilePath"
       continue
     fi
